+++ cypress/helpers/open_in_app.js
@@ -1,5 +1,5 @@
-import { dashboards } from '../../assets/backends/index.js'
-import { clickItemMenuButton } from '../../elements/dashboardItem.js'
+import { dashboards } from '../assets/backends/index.js'
+import { clickItemMenuButton } from '../elements/dashboardItem.js'
 
 // these tests being run on the "Delivery" dashboard
 const chartItemUid = dashboards.Delivery.items.chart.itemUid
