+++ cypress/helpers/view_as.js
@@ -1,12 +1,12 @@
-import { dashboards } from '../../assets/backends/index.js'
+import { dashboards } from '../assets/backends/index.js'
 import {
     chartClass,
     mapClass,
     tableClass,
     clickItemMenuButton,
     getDashboardItem,
-} from '../../elements/dashboardItem.js'
-import { EXTENDED_TIMEOUT } from '../../support/utils.js'
+} from '../elements/dashboardItem.js'
+import { EXTENDED_TIMEOUT } from '../support/utils.js'
 
 // these tests being run on the "Delivery" dashboard
 const chartItemUid = dashboards.Delivery.items.chart.itemUid
